chore: push latest

+++ TransparentCloudServerProxy.WebDashboard/Controllers/ProxyApiController.cs
@@ -1,8 +1,13 @@
+using System.Security.Claims;
+
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 
 using TransparentCloudServerProxy.ProxyBackend;
 using TransparentCloudServerProxy.WebDashboard.Services.Interfaces;
+using TransparentCloudServerProxy.WebDashboard.SqlDb;
+using TransparentCloudServerProxy.WebDashboard.SqlDb.Models;
 
 namespace TransparentCloudServerProxy.WebDashboard.Controllers {
     [ApiController]
@@ -11,10 +16,12 @@ namespace TransparentCloudServerProxy.WebDashboard.Controllers {
     public class ProxyApiController : ControllerBase {
         private readonly ILogger<ProxyApiController> _logger;
         private readonly IProxyService _proxyService;
+        private readonly IDbContextFactory<WebDashboardDbContext> _dbContextFactory;
 
-        public ProxyApiController(ILogger<ProxyApiController> logger, IProxyService proxyService) {
+        public ProxyApiController(ILogger<ProxyApiController> logger, IProxyService proxyService, IDbContextFactory<WebDashboardDbContext> dbContextFactory) {
             _logger = logger;
             _proxyService = proxyService;
+            _dbContextFactory = dbContextFactory;
         }
 
         private static IActionResult BadRequestError(Exception e) {
@@ -24,11 +31,39 @@ namespace TransparentCloudServerProxy.WebDashboard.Controllers {
                 Content = e.Message.ToString()
             };
         }
+        private async Task<ProxyUser?> GetCurrentUserAsync(WebDashboardDbContext dbContext) {
+            if (User.Identity is not ClaimsIdentity identity) {
+                return null;
+            }
+
+            var idClaim = identity.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (string.IsNullOrEmpty(idClaim) || !Guid.TryParse(idClaim, out var currentUserId)) {
+                return null;
+            }
+
+            var currentUser = await dbContext.Users
+                .Include(u => u.UserSavedProxies)
+                .SingleOrDefaultAsync(u => u.Id == currentUserId);
+
+            return currentUser;
+        }
+
 
         [HttpGet(nameof(GetProxies))]
-        public IActionResult GetProxies() {
+        public async Task<IActionResult> GetProxies() {
             try {
-                return Ok(_proxyService.GetProxies());
+                using var dbContext = _dbContextFactory.CreateDbContext();
+                var currentUser = await GetCurrentUserAsync(dbContext);
+                if (currentUser is null) {
+                    return BadRequest();
+                }
+
+                if (currentUser.Admin) {
+                    return Ok(_proxyService.GetProxies());
+                }
+                else {
+                    return Ok(currentUser.UserSavedProxies.Select(i => i.GetProxy()).ToArray());
+                }
             }
             catch (Exception e) {
                 return BadRequestError(e);
@@ -36,9 +71,22 @@ namespace TransparentCloudServerProxy.WebDashboard.Controllers {
         }
 
         [HttpPost(nameof(StartAllProxies))]
-        public IActionResult StartAllProxies() {
+        public async Task<IActionResult> StartAllProxies() {
             try {
-                _proxyService.StartAllProxies();
+                using var dbContext = _dbContextFactory.CreateDbContext();
+                var currentUser = await GetCurrentUserAsync(dbContext);
+                if (currentUser is null) {
+                    return BadRequest();
+                }
+
+                foreach (var savedProxy in currentUser.UserSavedProxies) {
+                    var proxy = savedProxy.GetProxy();
+                    if (proxy is null) {
+                        continue;
+                    }
+
+                    _proxyService.StartProxy(proxy);
+                }
             }
             catch (Exception e) {
                 return BadRequestError(e);
@@ -75,8 +123,18 @@ namespace TransparentCloudServerProxy.WebDashboard.Controllers {
         }
 
         [HttpPost(nameof(AddProxy))]
-        public IActionResult AddProxy([FromBody] Proxy proxy) {
+        public async Task<IActionResult> AddProxy([FromBody] Proxy proxy) {
             try {
+                using var dbContext = _dbContextFactory.CreateDbContext();
+                var currentUser = await GetCurrentUserAsync(dbContext);
+                if (currentUser is null) {
+                    return BadRequest();
+                }
+
+                var savedProxy = new SavedProxy(proxy, currentUser.Id);
+                dbContext.Proxies.Add(savedProxy);
+                dbContext.SaveChanges();
+
                 _proxyService.AddProxyEntry(proxy);
             }
             catch (Exception e) {
@@ -87,9 +145,19 @@ namespace TransparentCloudServerProxy.WebDashboard.Controllers {
         }
 
         [HttpPost(nameof(RemoveProxy))]
-        public IActionResult RemoveProxy([FromBody] Proxy proxy) {
+        public async Task<IActionResult> RemoveProxy([FromBody] Proxy proxy) {
             try {
+                using var dbContext = _dbContextFactory.CreateDbContext();
+                var currentUser = await GetCurrentUserAsync(dbContext);
+                if (currentUser is null) {
+                    return BadRequest();
+                }
+
+                var savedProxy = new SavedProxy(proxy, currentUser.Id);
+                var existingRule = currentUser.UserSavedProxies.Single(i => i.Id == savedProxy.Id);
+
                 _proxyService.RemoveProxyEntry(proxy);
+                currentUser.UserSavedProxies.Remove(existingRule);
             }
             catch (Exception e) {
                 return BadRequestError(e);
